Crate rs_graph

A library for basic graph data structures and algorithms.

Re-exports

pub use self::traits::Digraph;

pub use self::traits::Graph;

pub use self::traits::Network;

pub use self::traits::IndexDigraph;

pub use self::traits::IndexGraph;

pub use self::traits::IndexNetwork;

pub use self::traits::NumberedDigraph;

pub use self::traits::NumberedGraph;

pub use self::traits::NumberedNetwork;

pub use self::reversedigraph::reverse;

pub use self::reversedigraph::ReverseDigraph;

pub use crate::builder::Buildable;

pub use crate::builder::Builder;

pub use self::linkedlistgraph::LinkedListGraph;

pub use self::attributes::AttributedGraph;

pub use self::attributes::BiEdgeAttributes;

pub use self::attributes::EdgeAttributes;

pub use self::attributes::NodeAttributes;

pub use crate::vec::BiEdgeSlice;

pub use crate::vec::EdgeSlice;

pub use crate::vec::NodeSlice;

pub use crate::vec::BiEdgeVec;

pub use crate::vec::EdgeVec;

pub use crate::vec::NodeVec;

pub use crate::vec::IndexBiEdgeSlice;

pub use crate::vec::IndexEdgeSlice;

pub use crate::vec::IndexNodeSlice;

pub use crate::vec::IndexBiEdgeVec;

pub use crate::vec::IndexEdgeVec;

pub use crate::vec::IndexNodeVec;

Modules

adjacencies
Abstraction of neighboring edges.

algorithms
General algorithms working on graphs.

attributed	[
Deprecated
]
Extend a graph with attributes.

attributes
Extend a graph with attributes.

branching
Compute a maximum weight branching.

builder
Traits for constructing graphs.

classes
Graph classes Some common graph classes.

collections

draw
Visualizing graphs.

filtered
This is a graph adaptor suppressing some edges.

linkedlistgraph
A linked-list based graph implementation.

maxflow
Maximum Network Flow algorithms.

mst
Minimum spanning tree algorithms.

reversedigraph
Reverse the direction of the edges of a digraph.

search
Graph search algorithms.

shortestpath
Shortest path algorithms.

string

traits
Traits for graph data structures.

vec
This module provides vectors that can be indexed by nodes, edges or arcs.

Macros

biedgevec
Create a biedge vector associated with a LinkedListGraph.

edgevec
Create an edge vector associated with a LinkedListGraph.

idxbiedgevec
Create a biedge vector associated with a network.

idxedgevec
Create an edge vector associated with a graph.

idxnodevec
Create a node vector associated with a graph.

nodevec
Create a node vector associated with a LinkedListGraph.

Traits

EdgeMap
Trait for maps assigning values to edges of a graph.

EdgeMapMut
Trait for mutable maps assigning values to edges of a graph.

NodeMap
Trait for maps assigning values to nodes of a graph.

NodeMapMut
Trait for mutable maps assigning values to nodes of a graph.

Type Definitions

Net
The default graph type.