Module rs_graph::attributed

Deprecated since 0.17.0:
use rs-graph-derive crate instead

Extend a graph with attributes.

Sometimes one needs additional attributes associated with the nodes or edges of a graph. The simplest way is to store them in a NodeVec or EdgeVec. If these attributes are tightly related to the graph, it is convenient to store the graph together with the attributes in a single struct, e.g.

use rs_graph::LinkedListGraph;

struct MyGraph {
    graph: LinkedListGraph,
    balances: Vec<f64>,
    bounds: Vec<f64>,
}

The problem is that MyGraph itself does not implement the graph traits Graph, Digraph and so on, and can therefore not used directly for methods that need values of this type.

The struct Attributed in this module can be used to automate the tedious task of implementing the traits.

For a different approach, see the rs-graph-derive crate.

Structs

Attributed	[
Deprecated
]
Wrapper to attach attributes with a graph.

AttributedAttrs	[
Deprecated
]
The default implementation of attributes.

AttributedBuilder	[
Deprecated
]
Default builder for an attributes graph.

GraphAttrs	[
Deprecated
]
Accessor for graph attributes.

Traits

AttributedGraph	[
Deprecated
]
An attributed graph.

AttributedNetwork	[
Deprecated
]
An attributed network.

Attributes	[
Deprecated
]
This trait provides (mutable) access to the attributes of an attributed graph.

NetworkAttributes	[
Deprecated
]
This trait provides (mutable) access to the attributes of a network.