Module rs_graph::search

Graph search algorithms.

This module contains several standard search algorithms on graphs. These are algorithms that visit the nodes of a graph in some specific order.

All search algorithms are implemented as iterators. The iterators produce a sequence of nodes (together with algorithm specific additional information) in the order in which they are visited by the particular search strategy. This allows search algorithms to be used on infinite graphs (represented by implementors of [Adjacencies][crate::adjacencies::Adjacencies]), in which case the iterator is infinite.

Modules

astar
A* search.

bfs
Breadth-first-search.

biastar
Bidirectional A*-search.

dfs
Depth-first-search.

Functions

path_from_incomings
Compute a path from a map of incoming edges for each node.