Function rs_graph::shortestpath::floydwarshall::undirected

pub fn undirected<'a, G, Ws, W>(
    g: &'a G, 
    weights: Ws
) -> Vec<Vec<Option<(W, G::Node)>>> where
    G: IndexGraph<'a>,
    Ws: EdgeMap<'a, G, W>,
    W: NumAssign + Ord + Copy + Bounded, 

Solve the All-Pairs-Shortest-Path-Problem with the algorithm of Floyd and Warshall on an undirected graph.

Returns a 2D vector with entries (dist, pred) for each pair of nodes where dist is the length of the shortest path and pred is the predecessor of the last node.