Trait rs_graph::traits::Directed

pub trait Directed<'a>: Undirected<'a> {
    type OutEdgeIter: 'a + Iterator<Item = (Self::Edge, Self::Node)>;
    type InEdgeIter: 'a + Iterator<Item = (Self::Edge, Self::Node)>;
    fn src(&'a self, e: Self::Edge) -> Self::Node;

    fn snk(&'a self, e: Self::Edge) -> Self::Node;

    fn outedges(&'a self, u: Self::Node) -> Self::OutEdgeIter;

    fn inedges(&'a self, u: Self::Node) -> Self::InEdgeIter;

    fn outgoing(&'a self) -> OutEdges<'a, Self>
    where
        Self: Sized,
    { ... }

    fn incoming(&'a self) -> InEdges<'a, Self>
    where
        Self: Sized,
    { ... }
}

A graph with list access to directed incident edges.

Note that each directed graph is also an undirected graph by simply ignoring the direction of each edge. Hence, each type implementing Directed must also implement Undirected.

This trait adds a few additional methods to explicitely access the direction information of an edge. In particular, the direction information can be used in the following ways:

The src and snk methods return the source and sink nodes of an edge.

The iterators outedges and inedges iterate only over edges leaving or entering a certain node, respectively.

Associated Types

type OutEdgeIter: 'a + Iterator<Item = (Self::Edge, Self::Node)>

Type of an iterator over the forward edges leaving a node.

type InEdgeIter: 'a + Iterator<Item = (Self::Edge, Self::Node)>

Type of an iterator over the backward edges entering a node.

Loading content...
Required methods

fn src(&'a self, e: Self::Edge) -> Self::Node

Return the source node of an edge.

fn snk(&'a self, e: Self::Edge) -> Self::Node

Return the sink node of an edge.

fn outedges(&'a self, u: Self::Node) -> Self::OutEdgeIter

Return an iterator over the outgoing edges of a node.

The iterator returns only forward edges.

fn inedges(&'a self, u: Self::Node) -> Self::InEdgeIter

Return an iterator over the incoming edges of a node.

The iterator returns only backward edges.

Loading content...
Provided methods

fn outgoing(&'a self) -> OutEdges<'a, Self> where
Self: Sized,

Return access to the outgoing arcs via an Adjacencies trait.

This is the same as calling OutEdges(&g) on the graph.

fn incoming(&'a self) -> InEdges<'a, Self> where
Self: Sized,

Return access to the incoming arcs via an Adjacencies trait.

This is the same as calling InEdges(&g) on the graph.

Loading content...