Trait rs_graph::traits::GraphSize

pub trait GraphSize<'a>: GraphType<'a> {
    type NodeIter: 'a + Iterator<Item = Self::Node>;
    type EdgeIter: 'a + Iterator<Item = Self::Edge>;
    fn num_nodes(&self) -> usize;

    fn num_edges(&self) -> usize;

    fn nodes(&'a self) -> Self::NodeIter;

    fn edges(&'a self) -> Self::EdgeIter;
}

A (finite) graph with a known number of nodes and edges.

Finite graphs also provide access to the list of all nodes and edges.

Associated Types

type NodeIter: 'a + Iterator<Item = Self::Node>

Type of an iterator over all nodes.

type EdgeIter: 'a + Iterator<Item = Self::Edge>

Type of an iterator over all edges.

Loading content...
Required methods

fn num_nodes(&self) -> usize

Return the number of nodes in the graph.

fn num_edges(&self) -> usize

Return the number of edges in the graph.

fn nodes(&'a self) -> Self::NodeIter

Return an iterator over all nodes.

fn edges(&'a self) -> Self::EdgeIter

Return an iterator over all edges.

This iterator traverses only the forward edges.

Loading content...