Trait rs_graph::traits::Undirected

pub trait Undirected<'a>: GraphType<'a> {
    type NeighIter: 'a + Iterator<Item = (Self::Edge, Self::Node)>;
    fn enodes(&'a self, e: Self::Edge) -> (Self::Node, Self::Node);

    fn neighs(&'a self, u: Self::Node) -> Self::NeighIter;

    fn neighbors(&'a self) -> Neighbors<'a, Self>
    where
        Self: Sized,
    { ... }
}

A graph with list access to undirected incident edges.

Associated Types

type NeighIter: 'a + Iterator<Item = (Self::Edge, Self::Node)>

Type of an iterator over all incident edges.

Loading content...
Required methods

fn enodes(&'a self, e: Self::Edge) -> (Self::Node, Self::Node)

Return the nodes connected by an edge.

The order of the nodes is undefined.

fn neighs(&'a self, u: Self::Node) -> Self::NeighIter

Return an iterator over the edges adjacent to some node.

This iterator traverses only the forward edges.

Loading content...
Provided methods

fn neighbors(&'a self) -> Neighbors<'a, Self> where
Self: Sized,

Return access to the neighbors via an Adjacencies trait.

This is the same as calling Neighbors(&g) on the graph.

Loading content...