Module rs_graph::vec

This module provides vectors that can be indexed by nodes, edges or arcs.

Structs

BiEdgeIndexer
Indexer for biedges of a network.

EdgeIndexer
Indexer for edges.

IndexItemSlice
A borrowed slice indexed by items of a graph.

IndexItemSliceMut
A mutable borrowed slice indexed by items of a graph.

IndexItemVec
An owned vector indexed by items of a graph.

ItemSlice

ItemVec
New type wrapper around Vec<T> to be indexed by nodes and edges.

NodeIndexer
Indexer for nodes.

Traits

Indexer
A mapper of graph items to indices.

NumItems
Returns the number of items in a graph.

NumberedIndexer
This trait maps nodes or edges to an index.

Type Definitions

BiEdgeSlice
Vector of values for bidirected edges of a graph.

BiEdgeVec
Vector of values for bidirected edges of a graph.

EdgeSlice
Vector of values for edges of a graph.

EdgeVec
Vector of values for edges of a graph.

IndexBiEdgeSlice
Slice of a biedge indexed vector.

IndexBiEdgeSliceMut
Mutable slice of a biedge indexed vector.

IndexBiEdgeVec
biedge indexed vector.

IndexEdgeSlice
Slice of a edge indexed vector.

IndexEdgeSliceMut
Mutable slice of a edge indexed vector.

IndexEdgeVec
Edge indexed vector.

IndexNodeSlice
Slice of a node indexed vector.

IndexNodeSliceMut
Mutable slice of a node indexed vector.

IndexNodeVec
Node indexed vector.

NodeSlice
Vector of values for nodes of a graph.

NodeVec
Vector of values for nodes of a graph.